+++ CHANGES.rst
@@ -5,4 +5,8 @@ Changelog of nens-graph
 0.1 (unreleased)
 ----------------
 
-- Initial library skeleton created by nensskel.  [your name]
+- Initial library skeleton created by nensskel.
+
+- Copied orinal lizard-map.adapter.graph to basic.
+
+- Added common graph class and deltaportaal-specific river graph